# src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: Financial Management API
  description: |
    A comprehensive REST API for financial management system that handles users, accounts, and transactions.
    
    ## Features
    - User management (CRUD operations)
    - Account management with different account types
    - Transaction processing and history
    - Real-time balance updates
    - Transaction status tracking
    
    ## Authentication
    Currently, this API does not require authentication. In production, implement proper security measures.
    
    ## Rate Limiting
    API calls are limited to 1000 requests per hour per IP address.
    
  version: 1.0.0
  contact:
    name: Financial API Support
    email: support@financial.com
    url: https://financial.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api
    description: Development server
  - url: https://api.financial.com/v1
    description: Production server
...
paths:
...
  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get account by ID
      description: Retrieve a specific account by its ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          description: Account ID
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Account:
      type: object
      required:
        - id
        - userId
        - accountNumber
        - accountType
        - balance
        - currency
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the account
          example: 1
        userId:
          type: integer
          format: int64
          description: ID of the user who owns this account
          example: 1
        accountNumber:
          type: string
          maxLength: 20
          description: Unique account number
          example: "ACC001"
        accountType:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
          format: decimal
          description: Current account balance
          example: 1500.00
        currency:
          type: string
          maxLength: 3
          description: Account currency code
          example: "USD"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Account last update timestamp
          example: "2023-01-01T00:00:00Z"
...
    AccountType:
      type: string
      enum:
        - CHECKING
        - SAVINGS
        - CREDIT
        - INVESTMENT
      description: Type of account
      example: "CHECKING"
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2023-01-01T00:00:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Validation failed"
        path:
          type: string
          description: Request path
          example: "/api/users"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []